+++ src/components/Battery/Battery.tsx
@@ -6,7 +6,6 @@
 
 import React, { useEffect, useRef, useState } from 'react';
 import { useSelector } from 'react-redux';
-import { Toggle } from '@nordicsemiconductor/pc-nrfconnect-shared';
 
 import { DocumentationTooltip } from '../../features/pmicControl/npm/documentation/documentation';
 import {
@@ -14,9 +13,8 @@ import {
     PmicChargingState,
 } from '../../features/pmicControl/npm/types';
 import {
-    getFuelGaugeSettings,
+    getFuelGaugeEnabled,
     getLatestAdcSample,
-    getNpmDevice,
     getPmicChargingState,
     isBatteryConnected,
 } from '../../features/pmicControl/pmicControlSlice';
@@ -148,11 +146,10 @@ export default ({ disabled }: BatteryProperties) => {
     const [iconSize, setIconSize] = useState(0);
     const iconWrapper = useRef<HTMLDivElement | null>(null);
 
-    const fuelGaugeSettings = useSelector(getFuelGaugeSettings);
+    const fuelGaugeEnabled = useSelector(getFuelGaugeEnabled);
     const pmicChargingState = useSelector(getPmicChargingState);
     const latestAdcSample = useSelector(getLatestAdcSample);
     const batteryConnected = useSelector(isBatteryConnected);
-    const npmDevice = useSelector(getNpmDevice);
 
     useEffect(() => {
         const newIconSize = (iconWrapper.current?.clientHeight ?? 20) * 0.9;
@@ -162,68 +159,44 @@ export default ({ disabled }: BatteryProperties) => {
     const charging = isCharging(pmicChargingState);
 
     return (
-        <div className="tw-flex tw-flex-col tw-gap-4 ">
-            {npmDevice?.fuelGaugeModule?.set.discardPosiiveDeltaZ !==
-                undefined && (
-                <Toggle
-                    label={
-                        <DocumentationTooltip
-                            card="battery"
-                            item="DiscardPositiveDeltaZ"
-                        >
-                            <span>Allow state-of-charge to increase</span>
-                        </DocumentationTooltip>
-                    }
-                    isToggled={!fuelGaugeSettings.discardPosiiveDeltaZ}
-                    disabled={disabled}
-                    onToggle={v =>
-                        npmDevice?.fuelGaugeModule?.set.discardPosiiveDeltaZ?.(
-                            v
-                        )
-                    }
-                />
-            )}
-            <div className={`battery ${disabled ? 'disabled' : ''}`}>
-                <div className="battery-wrapper">
-                    <div className="battery-graphic-wrapper">
-                        <div className="battery-graphic-wrapper-wrapper">
-                            <div className="battery-graphic">
-                                <div className="battery-nipple" />
-                                <div className="battery">
-                                    <div
-                                        className={`gauge ${
-                                            charging ? 'animated' : ''
-                                        } ${charging ? 'charging' : ''}`}
-                                        style={{
-                                            height: `calc(${
-                                                fuelGaugeSettings.enabled &&
-                                                batteryConnected &&
-                                                !Number.isNaN(
-                                                    latestAdcSample?.soc ?? 0
-                                                )
-                                                    ? latestAdcSample?.soc ?? 0
-                                                    : 0
-                                            }% + 2px)`,
-                                        }}
-                                    />
-                                </div>
+        <div className={`battery ${disabled ? 'disabled' : ''}`}>
+            <div className="battery-wrapper">
+                <div className="battery-graphic-wrapper">
+                    <div className="battery-graphic-wrapper-wrapper">
+                        <div className="battery-graphic">
+                            <div className="battery-nipple" />
+                            <div className="battery">
                                 <div
-                                    className={`state-missing ${
-                                        !batteryConnected ? '' : 'hidden'
-                                    }`}
-                                />
-                                <BatterIcon
-                                    pmicChargingState={pmicChargingState}
+                                    className={`gauge ${
+                                        charging ? 'animated' : ''
+                                    } ${charging ? 'charging' : ''}`}
+                                    style={{
+                                        height: `calc(${
+                                            fuelGaugeEnabled &&
+                                            batteryConnected &&
+                                            !Number.isNaN(
+                                                latestAdcSample?.soc ?? 0
+                                            )
+                                                ? latestAdcSample?.soc ?? 0
+                                                : 0
+                                        }% + 2px)`,
+                                    }}
                                 />
                             </div>
+                            <div
+                                className={`state-missing ${
+                                    !batteryConnected ? '' : 'hidden'
+                                }`}
+                            />
+                            <BatterIcon pmicChargingState={pmicChargingState} />
                         </div>
                     </div>
-                    <SideText
-                        batteryConnected={batteryConnected}
-                        latestAdcSample={latestAdcSample}
-                        fuelGaugeEnabled={fuelGaugeSettings.enabled}
-                    />
                 </div>
+                <SideText
+                    batteryConnected={batteryConnected}
+                    latestAdcSample={latestAdcSample}
+                    fuelGaugeEnabled={fuelGaugeEnabled}
+                />
             </div>
         </div>
     );
